chore:  restruct pallet burning
